update-routers
