+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getBuId = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const removeById = require("./removeById");
+
+module.exports = {
+    getAll,
+    getBuId,
+    add,
+    updateById,
+    removeById,
+}
